+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-## [2.0.0] - 2026-08-11
+## [2.0.1] - 2026-08-13
 
-### AI Interview Feature
-Added AI voice interview feature. Users can now conduct private voice interview practice using AI built into Torcbot. Users will be able to provide the job description as well as the type of interview they would like to go through, i.e., behavioral, technical, conversational or case-based. The user can also provide the company and the bot will attempt to conduct the interview in that companies style if data is available for how the provided company conducts their interviews.
+### Interview feature updates
+- Added a transcript of the questions that were asked and the answers that were provided to the DM response after the interview has concluded so that the interviewee can see the questions and their responses. 
+- Added the ability to select a preferred language to conduct the interview in for global communities.